requirements.txt
numpy==1.19.2
pandas==1.1.3
streamlit==0.76.0
Pillow==8.0.1